Create README.md
This is my master thesis dual active bridge DC-DC converter. The principle of operation of the DAB converter is as follows. There is a DC power source. It is converted to alternating current using an inverter. This AC voltage will increased or decreased using a transformer. There is a capacitor to filter the DC output voltage. So it's basically going to get DC to DC. It gives a transformer in which the voltage can be adjusted by appropriate choice of the turn ratio can be increased or decreased.

Short explanation about code : 
-The frequency interval is between 25kHz-50kHz, I used switch case statement for that, if else can also be used.
-The map command syntax is as follows:
map(value, fromLow, fromHigh, toLow, toHigh)
It is a command that, depending on the magnitude of the value, takes a value between 25 and 50 kHz (25,30,35,40,45,50) to a value between 640 and 320. For example The Value 40 corresponds to 400(16MHz/40kHZ = 400), or value 50 corresponds to 320(16MHz/50kHz=320)
-Serial.begin(9600); Command starts the serial console.
pinMode(); Command sets specified pin as input or output. 
-GTCCR = (1<<TSM)|(1<<PSRASY)|(1<<PSRSYNC); Command stops all timers
-The phase shift can be changed through the connected potentiometer. A mean value is calculated before transmission as the input is susceptible to electromagnetic interference. As in code, the potentiometer value is taken and added 10 times and then through Divided by 10 to get the mean value.